Refactor HarborFi adapter: extract addresses and clarify dynamic discovery status
- Moved hardcoded MINTERS and GENESIS_CONTRACTS to `addresses.js` for better maintainability.
- Updated TODO in `index.js` to reflect that dynamic discovery via factory events was attempted but failed due to missing on-chain events.
- Verified TVL remains consistent.
